+++ Runtime/Runner/IQosRunner.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Unity.Services.Qos.Internal;
 using Unity.Services.Qos.Models;
 
 namespace Unity.Services.Qos.Runner
@@ -9,5 +8,8 @@ namespace Unity.Services.Qos.Runner
     {
         // Returning a List for simpler sorting (IList doesn't have a Sort method)
         Task<List<Internal.QosResult>> MeasureQosAsync(IList<QosServer> servers);
+
+        // Returning a List for simpler sorting (IList doesn't have a Sort method)
+        Task<List<QosAnnotatedResult>> MeasureQosAsync(IList<QosServiceServer> servers);
     }
 }